ref: move formatting code into new file
